requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-mistralai
openai
mistralai
chromadb
tiktoken
sentence-transformers
pypdf
pdfplumber
python-docx
unstructured
beautifulsoup4
lxml
nltk
python-dotenv
fastapi
uvicorn
streamlit
numpy
pandas
scikit-learn
tenacity
rank-bm25
transformers
torch
loguru
aiohttp